Add/Change #193
MessageProcessor::call_RAY_MPI_TAG_GET_PATH_LENGTH(Message*): Assertion `m_fusionData->m_FUSION_identifier_map.count(id)>0' failed

+++ code/ArrayOfReads.cpp
@@ -79,7 +79,7 @@ void ArrayOfReads::push_back(Read*a){
 	#endif
 }
 
-int ArrayOfReads::size(){
+uint64_t ArrayOfReads::size(){
 	return m_elements;
 }
